Implement timeout for pcgamesn
An asset from the pcgamesn.com URL (https://kumo.network-n.com/dist/css/pcgamesn.css) runs for ~175 seconds before failing and it blocks loading the rest of the page. This causes the Got function to timeout, since it takes too long to load the page. As a workaround, an extended timeout was implemented so that the Got waits at least 5 minutes before timing itself out.